ciftlikpyqt/requirements.txt
PyQt6==6.6.1
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.9
python-dateutil==2.8.2 